third_party/WebKit/LayoutTests/external/wpt/feature-policy/experimental-features/sync-script.tentative.https.sub-expected.txt

This is a testharness.js-based test.
FAIL inline parser blocking script should be blocked assert_equals: expected (boolean) true but got (undefined) undefined
Harness: the test ran to completion.
